requirements.txt
Flask==3.0.0
python-dotenv==1.0.1
pymongo==4.7.3
bcrypt==4.1.2
google-genai>=0.5.0
Flask-Mail>=0.9.1
openai>=1.0.0
groq>=0.4.0
